requirements.txt
absl-py==0.13.0
angles==2.0
asn1crypto==0.24.0
astunparse==1.6.3
bcrypt==3.2.0
bondpy==1.8.5
Brlapi==0.6.6
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.2
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
certifi==2018.1.18
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
colorama==0.3.7
command-not-found==0.3
cryptography==3.4.7
cupshelpers==1.0
cv-bridge==1.13.0
dataclasses==0.8
defer==1.0.6
diagnostic-updater==1.9.7
distro-info===0.18ubuntu0.18.04.1
dlib==19.22.0
Flask==2.0.1
flatbuffers==1.12
gast==0.4.0
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
google-api-core==2.2.2
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-cloud-texttospeech==2.7.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.41.1
grpcio-status==1.41.1
gTTS==2.2.3
h5py==3.1.0
httplib2==0.9.2
idna==2.6
image-geometry==1.13.0
importlib-metadata==4.6.1
imutils==0.5.4
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
joint-state-publisher==1.12.15
jwt==1.2.0
kazam==1.4.5
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
libsvm==3.23.0.4
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.4
MarkupSafe==2.0.1
message-filters==1.14.11
msgpack==1.0.2
netifaces==0.10.4
nltk==3.6.2
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.1
olefile==0.45.1
opencv-python==4.5.3.56
opt-einsum==3.3.0
pexpect==4.2.1
Pillow==5.1.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.0
pytesseract==0.3.8
python-apt==1.6.5+ubuntu0.6
python-dateutil==2.6.1
python-debian==0.1.32
python-qt-binding==0.4.3
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-py-common==0.4.2
regex==2021.7.6
reportlab==3.4.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.11
roslaunch==1.14.11
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.11
rosmake==1.14.9
rosmaster==1.14.11
rosparam==1.14.11
rospy==1.14.11
rostest==1.14.11
rosunit==1.14.9
rqt-gui==0.5.2
rqt-gui-py==0.5.2
rqt-py-console==0.4.8
rqt-runtime-monitor==0.5.7
rqt-shell==0.4.9
rqt-top==0.4.8
rqt-web==0.4.8
rsa==4.7.2
scipy==1.5.4
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sensor-msgs==1.12.8
simplejson==3.13.2
six==1.15.0
smach==2.0.1
smclib==1.8.5
system-service==0.3
systemd-python==234
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.5.0
tensorflow-gpu==2.5.0
termcolor==1.1.0
tesseract==0.1.3
tf-slim==1.1.0
tf2-py==0.6.5
timer==0.1.2
topic-tools==1.14.11
tqdm==4.61.2
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.5
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==2.0.1
wrapt==1.12.1
xacro==1.13.11
xkit==0.0.0
zipp==3.5.0
zope.interface==4.3.2
